Add job walk and a time comparison test
This doesn't account for the fact that a job can be added to the
front of the spoke while we are looking.
Maybe lock a spoke during walk and job add with a common mutex?
You can either consume a new job or add a new one (which might add
a new one up front) - or if a new job is coming in whose trigger
time is in the past, bypass it and add it to a global `past-spoke`
and always walk that spoke first.

Once we have multiple spokes, we can make an `iterator` of spokes that
we can `walk` on.